requirements.txt
Flask==2.3.2
Pillow==10.0.0
summa==1.2.0
nltk==3.8.1
